requirements.txt
django==3.2.7
paddlepaddle==3.0.0
paddleocr==3.0.0
opencv-python==4.11.0.86
numpy>=1.21.0
pillow>=8.3.2
scikit-learn>=0.24.2
matplotlib>=3.4.3
seaborn>=0.11.2
wordcloud>=1.8.1
jieba>=0.42.1
python-docx>=0.8.11
PyPDF2>=2.0.0
requests>=2.26.0
beautifulsoup4>=4.9.3
tqdm>=4.62.3
pandas>=1.3.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
plotly==5.17.0 